+++ R/heading_tree.R
@@ -1,19 +1,18 @@
 #' This constructs a data frame of headings for displaying to the user
 #' via the CLI package.
 #'
-#' @param headings an object of class `xml_nodelist`
+#' @param headings a table of headings that contains headings with the text of
+#'   the heading, the level and the position
 #' @return a data frame with two columns:
 #'  - heading: the text of the heading with ATX header levels prepended
 #'  - children: a column of lists indicating the immediate children of the 
 #'    heading. This is used to display a fancy tree from the cli package. 
 #' @noRd
-heading_tree <- function(headings, lname = NULL, suffix = NULL) {
+heading_tree <- function(htab, lname = NULL, suffix = NULL) {
 
-  lname   <- if (is.null(lname)) "<LESSON>" else dQuote(lname)
-  hlevels <- as.integer(xml2::xml_attr(headings, "level"))
-  hnames  <- xml2::xml_text(headings)
-  hnames  <- c(lname, hnames)
-  hlevels <- c(1L, hlevels)
+  lname   <- if (is.null(lname)) "<EPISODE>" else dQuote(lname)
+  hnames  <- c(lname, htab$heading)
+  hlevels <- c(1L, htab$level)
   hlabels <- vapply(hlevels, switch, character(1),
     `1` = "#",
     `2` = "##",
@@ -23,7 +22,7 @@ heading_tree <- function(headings, lname = NULL, suffix = NULL) {
     `6` = "######" 
   )
   hlevels[1] <- 0L
-  hlabels[1] <- "# Lesson:"
+  hlabels[1] <- "# Episode:"
   hnames <- paste(hlabels, hnames)
   htree <- data.frame(
     heading  = hnames,
@@ -75,6 +74,7 @@ label_duplicates <- function(htree, cli = FALSE) {
   get_duplicated <- function(i) duplicated(i) | duplicated(i, fromLast = TRUE)
   the_clones <- lapply(htree$children, get_duplicated)
   has_twins <- vapply(the_clones, any, logical(1))
+  are_unique <- logical(length(the_clones)) | TRUE
 
   dtree <- htree$labels
   for (i in which(has_twins)) {
@@ -87,6 +87,7 @@ label_duplicates <- function(htree, cli = FALSE) {
       e = if (cli) cli::style_inverse("(duplicated)") else "(duplicated)",
       cli = FALSE
     )
+    are_unique <- are_unique & !the_duplicated
     # This part is needed to fix a feature in CLI that assumes unordered data
     # 
     # If there is a child of a duplicated node, CLI does not know where to put
@@ -104,7 +105,7 @@ label_duplicates <- function(htree, cli = FALSE) {
     }
   }
   htree$labels <- dtree
-  list(test = !any(has_twins), tree = htree)
+  list(test = are_unique, tree = htree)
 }